Fixes for the dreaded wmap loop. (For those interested, merging the checking and the compiling is a BAD idea unless you really only want run-time checking. I can explain more on Monday.)
Also fixed several other bugs yesterday, they're included in this
commit but I can't remember the details.

The parsing of types is still a little fragile. I spent a long time
figting with ocamlyacc, but didn't resolve it yet. For now, be careful
where you stick identifiers, and if you get a parse error, check out
the types section of parser.mly. It's pretty straightforward.