Finished porting normalization code and member checking. SIMPLE aconds now work for me.
I'm committing this now, so that others can get on with their work,
but it is still very much alpha. In particular, make sure that you
write down "obviously projectable" types: i.e., at each level in a
type, for each name n, the type under n should be the exact same type. So

      {!="X",*\(n,m)="X"} 

is OK but

      {!="X",*\(n,m)="Y"} 

is not. 

With this assumption, model checking for types is simple. Later on
(maybe now) we should do the more general thing that constructs the
bipartite graph and tries to find a perfect matching. It wouldn't be
that inefficient...