OK. Visualizer is usable now.
Error display is not very fancy. I had to turn off the tagging
(yellow syntax highlighting) that Malo added because the arithmetic is
wrong and it's causing GTK assertions to blow up at the C level. 

But basic usage is OK. 

Don't forget to pass a -I flag on the commandline, or you won't have
your Prelude :-)