Moved utility functions from harmony.ml to surveyor.ml
Cleaned up front end

Added a real "load" function that takes a name (like "abc.tmp:meta"),
locates it in the search path and uses the surveyor to select a
viewer. It fails (1) if it can't find a unique viewer or (2) no file
is found in the path, or (3) it can't be parsed using the viewer.