Here's the big commit. New parser, new compiler, ported lens library.
Everything in examples works EXCEPT for the very last sync test in
structuredtext. An explode is failing because it is presented with a
view with multiple children.

Rather than record all the changes, I think it's just easiest for us
to sync up in the AM. I plan on begin at Penn no later than 10am.

Still to do:
      - improve checking of recursive variables
          * automatic delaying of lenses

      - uncomment/port type normalizer, test for non-termination in
        type operations