After the discussion with Alan & Benjamin, ported the iCalendar files to Harmony4 and added them in a iCalendar subdirectory in src.
I also added a iCalendar directory in experimental and a test on a
ics files. There is no lens at work right now, just the reader and 
this test is just to check that everything is going well when the 
file is	converted into meta.

Now, be careful with making clean... because 95% of the time spent to
compile the whole Harmony is spent in iCalendarlex.ml (looking at the
file, it's no wonder !). I hesitated adding iCalendar/ or not to the
SUBDIRSCLEANONLY variable in src/Makefile, and I decided to put it for
the sake of coherence. If the compilation time bothers you, consider
removing it and typing make clean in src/iCalendar/ manually when
really needed.