Fixed the bug with the missing schemas in harmonize-addresses (and bookmarks, according to Alan).
The bug appeared because we switched from linking directly against the
object code to linking against harmony.cmxa. Here's how ocamlopt uses
libraries (from the manpage):

> Arguments  ending  in  .cmxa  are taken to be libraries of object code.
> Such a library packs in two files lib.cmxa and lib.a a  set  of  object
> files  (.cmx/.o  files).  Libraries are build with ocamlopt -a (see the
> description of the -a option below). The object files contained in  the
> library  are  linked  as  regular  .cmx files (see above), in the order
> specified when the library was built. The only difference is that if an
> object  file  contained  in a library is not referenced anywhere in the
> program, then it is not linked in.

Note that modules that are not referenced do not get linked
in. Harmony is a little funny because if you chase the modules from
the top level, (at least) viewers.cmx, prelude.cmx, and compiler.cmx
are not directly referenced. E.g., the compiler is only used to
compile dynamically-loaded .fcl files via a backpatch reference in the
registry; Focal modules (like prelude.src) expect Registry.get_library
to contain values for QIds like Native.Prelude.xfork, but do not
actually link against the OCaml module... and so on.

So if we're going to use an OCaml library to package up the Harmony
code for linking against by front-ends (and I'm not sure this is a
good idea, because of these complications), we need to be sure to
explicitly load these otherwise only implicitly-referenced modules. 

I've added [init ()] functions for Compiler and Prelude, and applied
those functions in Toplevel. This is enough to fix the bug with "make
test10" in examples/addresses.