prettier pretty printing.
Stephane: I removed your stuff from schemas.ml, because I couldn't make sense of it. Can you add it back in? Note that I changed Wildcards. They upper bound, which used to be an int option is now just a bool --- whether or not it's n bangs and a star, or just n bangs. make sense?