LOTS more progress on the web demos.
I got frustrated with the fragility and hackiness of the
makefile-based demos, so I finally just junked them in favor of a
pure-PHP solution.  (I'm a halfway decent PHP programmer now. :-)

I still don't have all this installed on a web server where you can
play with it, but if you look at examples/structuredtext/demos.php you
should get a clear idea what's going on.  

I still need to add a little more functionality so that people can
play with lenses in addition to synchronization (this is also needed
for some of Aaron's demos).  And I hope some CSS hacker can help make
it a little prettier.  :-) But basically it is now working very well.
I'm going to use it for the demo on Wednesday to the DB group.