Preliminary support for printing of actions containing spined lists.
There are still some significant shortcomings.  In particular, because
of the way synchronization of spined lists works, changing an element
in the middle of a spined list will be synchronized (which we
intended) and printed (which is a problem) as a replacement of the
entire tail of the list.

A smaller shortcoming is that I piggybacked the spined list printing
on the list printing in a simple way, so the outer brackets print as
[...] rather than [|...|].

Thoughts?

(If you want to play with it, use the Playground area of the demo.)