implemented "baking in" of focal libraries to the Harmony binary, along the lines that benjamin suggested.
baker.ml scans the list of files and/or directories given on the
command line, recursively searches for all the .fcl and .src files it
can find (choosing .src over .fcl where both files are defined and
clobbering sources with the same prefix arbitrarily), and emits code
to initialize a hashtable with a bool (to indicate if it is fcl or
src) and the contents of the file. 

when searching for a module, the registry checks the filesystem first
(using the search paths as before). if none is found, it checks the
Bakery.items hash. 

minimally tested, but compare these (make sure FOCALPATH is unset)

  harmony -check Prelude -verbose

  harmony -check Prelude -verbose -I path/to/lenses