Bookmark hacking.
We now handle the FEEDURL attribute in Mozilla Firefox bookmark
files. It's a little messy because the formats are messy. Safari
stores RSS feeds as regular old links with URIs of the form
"feed://..." whereas Firefox stores them under an attribute
FEEDURL="http://..." The way I decided to handle this is to rewrite
the Firefox-style entries to "feed://..." (using explode, implode, and
some conditionals. FUN!!)

Safari seems to keep some meta-data on RSS feeds. I haven't bothered
to worry about this. When I populated my Mozilla bookmarks with an RSS
feed and synchronized, it appeared in my Safari bookmark file even
though this section was not filled in... But if there are bugs, it's
something to keep in mind.