More cleaning up.
The distinction between .src and .fcl is gone. It made the registry
and code that walks the filesystem looking for Focal sources more
confusing than they had to be. All .fcl sources are "literate" and
start in SRC mode (escape to TEX with #{@} as usual). I left some
files with the .src extension if they are used to generate TeX and
non-Focal code (e.g., the parser and lexer in src/). Makefiles still
need some tidying.

Fixed web pages with suggestions from BCP, AS.