[Turned off buggy unit test in cycle list]
Added support for static checking of recursive lenses. This wasn't
hard, I just hadn't done it yet. The way I did it was kind of a hack:
keep SLens as a sort but add SRecLens of (exp * exp) as well. I didn't
want to code up new environments, so this way we just use the
compilation environment. To avoid evaluating those exps multiple
times, the parser desugars them into let-bindings of some fresh
variables and then uses those variables for the static check.

Fixed a bug in Schema.restrict: to calculate the restriction of a
schema to a set of names (and, simultaneously, the restriction to the
complement) we cannot just set some variables in the formula to
zero. Instead, we have to existentially quantify the variables in the
formula corresponding to names NOT included in the restriction. E.g.,
consider (List.T C). This schema is compiled in a semantic object that
looks like this:

  X = F[phi,basis] 

    where phi = (n0=1 & n1+n2+n3+n4+n5+n6=0) 
               | (n2=1 & n4=1 & n0+n1+n3+n5+n6=0)

    and basis = [ n0={nil}[{}], n1={nil}[T-{}],
                  n2={hd}[C], n3={hd}[T-C],
                  n4={tl}[X], n5={tl}[T-X],
                  n6=~{nil,hd,tl}[T] ]         

A semantic object representing the restriction of this schema to {tl}
can be obtained by:

  (1) making sure that {tl} is disjoint from any CoFinite locations
      (just do a refinement with the "true" basis that has every name
      in the restricted set, which gives us back a rewriting for the
      formula

  (2) throwing out any elements from the basis that do not fall in the
      set and existentially quantifying the variable corresponding to
      that element

  (3) adding back in a single CoFinite element containing the negation
      of the set, add a counting constraint set for this element set
      to 0. 

E.g., in this case we obtain

  X = F[phi',basis']
 
    where phi = EX x0,x1,x2,x3,x6.
                  ((x0=1 & x1+x2+x3+n0+n1+x6=0) 
                   | (x2=1 & n0=1 & x0+x1+x3+n1+x6=0))
                  & n2=0
    and basis' = 
        [ n0={tl}[X], n1={tl}[T-X],
          n2=~{tl}[T] ]

which is just 

    {} | { tl=X }

The restriction to the complement of a finite set is simpler since we
don't have to worry about adding back in a single CoFinite
element--after (1), we just include every CoFinite set in the basis. 

Prelude.xfork and Prelude.fork should typecheck correctly
now. Unfortunately, forks of maps, which we're using in place of
wmaps, don't have the correct type. Consider the following
declaration:

  schema C = List.T {a} in 
  let l : C <-> C = 
    fork {tl} (map (protect l)) id 
 
You'd think this would be correct? You'd be wrong :-) Well, you'd be
right, semantically, the lens does have that type; just not according
to the typing rules we wrote down. This was revealed by the static
checker.

First, we assume that l has the given type. Then, using the type for
fork, we split C by restricting it to {tl} and ~{tl}, yielding
schemas:

  C1 = {} | { tl=C }
  C2 = {nil} | { hd={a} }

Next we apply (map (protect l)) to the first branch and calculate
 
  A1 = {} | { tl=C }

using the hypothesis for the type of (protect l); we also check that
the names appearing in A1 are all included in {tl}.

Then we apply (id) to the second branch and calculate

  A2 = C2 = {nil} | { hd={a} }

and check that the names appearing in A2 are included in ~{tl}.

Finally we concatenate these schemas to obtain the type for l

     (C1+C2) <-> (A1+A2) 

i.e., ({nil} | {nil, tl=C} | {hd={a}} | {hd={a},tl=C}) 
  <-> ({nil} | {nil, tl=C} | {hd={a}} | {hd={a},tl=C})

Oops. Not quite what we intended eh?

Now we see why the type of wmap was useful: it let us reason that the
domain-set was stable :-/