I can't quite rememebr what's in this commit, but I think it's just small bug fixes and support in the parser for all three lens types. The syntax in this commit is:
  <~> for Bijective
  <=> for very well behaved
  <-> for well behaved.

Note that <=> looks like the total lens type we've been using in
LaTeX, but since the static checker is probably not going to track
totality (or should it? it wouldn't be hard to do so), I thought it
was best to use that symbol for something else.

Also, do we want to take Paul Taylor's suggestion and use an
asymmetric symbol for lens types? E.g., perhaps <<-> since there are
two arguments in the PutBack? This is kind of a trivial point, and
easy enough to change later, but if we're sure we want to do it, I'd
rather do it now before we annotate all our code and get used to
typing one string of symbols :-)