Fixed a bug in Safari lens that was breaking web demo.
Fixed a bug in Tree formatter introduced when we got rid of the Meta
viewer. Values in lists must now be formatted as [{a},{b},{c}] not
[a,b,c] because in the latter case, the compiler tries to lookup
variables named a, b, and c.

Peformance--both of compilation and of member testing--is still a
major issue, especially for Calendars. Alan, I may have to put back
the mk_fast_cat code :-(