Fixed bug in project that Conrad's changes turned up.
Added unit test to check simplified version of Conrad's example. 

Streamlining of oracle stuff (still not clear that this will be
useful--a miss on the oracle will trigger a long compilation--so the
oracle only saves you work if you avoid compiling a formula that would
otherwise not be used. Also, the string representation of oracle data
is dumb, if robust.).