Checkpoint tidying: removed fds hack.
Also, I forgot to log the following in my last commit, added a small
optimization to eliminate ANDed in True formulas, which were
mysteriously appearing. Here's the story: in lookup_cset, which calls
lookup_cvar, we take a list of "simple" type variable intersections
(tis) and differences (tds) and construct the type
              
  mk_diff (mk_isect tis) (mk_union tds)

  where mk_diff t1 t2 = mk_isect [t1; mk_neg t2]

The first line evaluates to

  mk_isect [mk_isect tis; mk_neg (mk_union tds)]

Now, if tds=[] then mk_union [] = V(False), and mk_neg V(False) =
V(True). It gets trickier... if the tis are all variables, then
mk_isect just does an isect on csets and mk_isect has a special case
for V(True). But if there are some non-variable types, then it has to
expose the formulas and bases, compute the refinements, and AND the
presburger formulas. Yikes.