An idea so simple, yet so sweet!
How would you tell that a schema is empty? You could do an expensive
co-inductive test to prove it. But if I've already shown that a
particular tree is a member of that schema, I bet I could give you an
answer faster. So obviously, every time we get a Yes answer from
member, we should add that type to the non_empties cache.

But wait, it gets better. Think abut the sync algorithm. First it
checks that each tree belongs to the schema. Then at each step it
projects some sub-schemas, and recurses, and calls dom_member on the
domains of the assembled results. Dom_member, in turn, uses the empty
test (to see if a domain is included in the domain-set of a schema, we
consider all the non-empty matching elements and construct a formula
that tests all possibilties of assigning names to elements--i.e.,
using some fresh variables and a sum). The key observation is that for
most of these empty tests, the answer is going to be no: we've just
checked that a *tree* belongs to the very same schema! So if we
remember the emptiness problems we've solved (in the negative) by a
counterexample, we can help out dom_member, and thus the synchronzier,
quite a bit.

In hindsight, it's kind of obvious, but I think this is a totally
sweet hack :-)

The USets are a little clunky and should be replaced by Hashsets or
something faster...