Optimization to fix the bug that Conrad noticed.
Too late to write a full explanation, but basically the first_step of
member was introducing a fresh variable for each child of the tree
being tested. For simple schemas, each child will be associated to a
single fresh variable, which must be 1... hence, we can safely use 1
instead of the fresh var. 

I believe this flattens out the performance curve. Conrad's 100
replica example was completing in 3.x seconds on our linux boxen.

Good night!