Crude "save" function for Conrad.
To make this change, I had to change the type of 'writers' in the
Surveyor from V.t -> unit to V.t -> string.  I suppose there are some
viewers (e.g., binary ones) that might work with the first type but
not the second, but we don't currently have any.  Until we do, let's
just stick with this type.

The basic functionality can be tested with
 
  test save "xml" (load "xml" "<a/>") = ?

Also fixed a bug in Value's formatter (closed Format'ting boxes), and
in the XML viewer (removed extra newlines).