Oops... I forgot something: the commit...
Beware that I finally changed something in the classical lenses, in the case of the use of "default". 
Before, for "default l def", we smashed the dictionnary from l.parse def in front of the incomming dictionnary. This has the (bad ?) effect that if these informations were not used at the creation point, they could appear in some other part of the putback.
In the case of streaming lenses, they are two types of dictionnary. The classic one, and the streaming one ("pointers" on the concrete file), and it was impossible to smash them directly. So I added a third type of dictionnary, containing two dictionnaries. A lookup in this dictionnary procedes in tow phases. First look in the first dictionnary, if found, return this value, if not, look in the second one. And now, for the "default" constructor, it allow us to remove the first dictionnary (wich is temporary), so that new informations don't flow out of this lens.