Adding quotient of lenses by canonizer. The behavior if it is not used at a top level need to be checked. I'm not 100% sure it would work if you build a qlens, and then use it as a chunk (inside <>)
A long commit because of indentation modification in rparser.mly, but not that much modification.

I choose >/ and /< for quotient. The "/" is for the quotient, and the > or < remind that on the left, we reduce the domain, and on the right, we enlarge the codomain. It's a suggestion, feel free to change !