Fixed these items from the TODO:   - Implicitly use BOOMPATH to "." when no environment variable is defined.   - Registry should search for filename, then lowercased filename.   - Figure out how to statically link to PCRE from GODI.   - Fix Makefile bugs in focal/extern/*
The issue with PCRE is subtle. The documentation for OCaml does not
agree with what happens on OS X. Essentially, ocamlopt does not
necessarily statically link to libraries, but just puts the same
information as conveyed by -ccopt parameters in the binary. The upshot
is that if you want to distribute binaries in a portable way, you need
to compile the *libraries* against the C-library .a files
directory. Luckily, PCRE seems to be the only library that is not
compiled this way, so I can just clobber the GODI installs of this
library on the two machines where I build OS X binaries with my own
version.