Little more tidying of simple CC lines.
Fixed Copyright stamp. We should talk about what we want it to read
project-wide. All the code in Boomerang was written by Alexandre +
me. Previously we've had BCP hold copyright (in case we want to
license to someone else). Since there are only three of us, we can fit
all three on the copyright line. If we grow to more contributors, I'd
vote to collapse to 1-2 and have a "CONTRIBUTORS" file for acks.