More syntax hacking...
 o Added "lens in C <-> A" and "string in R" in sorts /
   parameters. This actually looks quite nice.

 o Rationalized handling of infix operators, 
   and added a bunch more.

One more thought: we need some syntax for unit tests that raise
blame. This is a little clunky, but something like this...

  test f x y z = blame 17:35-50

(I don't like having to give the line numbers though. A crude but
prettier alternative: could give the name of a let-bound variable, and
check that its Info.t encompases the actual location that was blamed?)