Brzozowski is back!
I WAS just going to try to remove Crm.t (Character Range Maps) from
the NFA-based back-end, but I realized that it too has to build DFAs
to check splittability so I must be doing something wrong... and I
was!

 o Fixed bug in derivative of Rep[eated] expressions.

 o Added a fast path for splittability--if the derivative of (reverse
 t1) and t2 on every character is empty on at least one of the two,
 then t1 does not end in a character that starts a word in t2. (And
 we're better, because we don't actually have to check *every*
 character--can just check the characters in their character maps.)

 o Added a fast path for disjointness--if one regexp is a singleton,
 and we have its representative, then we can just use ordinary string
 matching. I don't think that this will be so important in general,
 but we have a disjoint union of several 100s of lenses in uniProtKB
 that hits this optimization.

  o Added code to make a tree of operators for unions. So if you write

     l1 | l2 | l3 | l4 | l5 | l6 | l7 | l8 

    the parser produces

     ( ( (l1 | l2) | (l3 | l4) ) | ( ( l5 | l6 ) | ( l7 | l8 ) ) )

Because of all of these optimizations, this isn't an apples-to-apples
comparison, but all unit tests now run in 42s (vs. 5-6min with NFAs)! 

There are still a few things that can be made faster, I think, but I'm
going to focus on collecting up some better unit tests.