Adding a few optimisation at different places:
- Bstring.dist now takes a limit option and returns a dist option
  If   the limit is None,
  Then it uses a heuristic to find a "good" limit and calls dist_limit
  Else it calls dist_limit with the min between limit and the result
    of the heuristic
  Note: if the limit was None, the result will always be: Some dist

- Bstring.dist_limit takes a limit and returns a dist option
  If   the distance is less than the limit
  Then the result is: Some dist
  Else it can be either Some dist or None

- Balign.diffy now uses a heuristic to find a limit and use this limit
  and other considerations to give a limit to the costs of put.  So it
  can use the preceding optimisation.

Bstring.dist_heurist works fine when the string are almost equal and
is linear in the sum of the size of the strings.

Bstring.dist_limit works fine when a good limit is given.  This
happens if the string are almost equal with the limit given by the
heuristic.  This also happens if the string are different with the
limit given by diffy.

The notion of limit should be extended to all alignments and passed by
when calling nested alignment in addition to the skeleton distance.