Grammar: - match ends with an optional sort - application operator ($): "f $ g x" is the same as "f(g(x))" - some changes to make the grammar more similar to OCaml - documentation - unit argument (), eg.: "let f () = ..." or "fun () -> ..." - etc
The sort in ECase now is optional (not sure if this was better than defining an ECaseUntyped).

When the type of the match is not defined, bcheck.ml/check_exp uses the first branch for the type. Match without type only works if the types are not refined (we are checking this)

Examples updated (using exception module and new grammar stuff)