Fix bprint.ml (ocaml exception because of a bad format when debugging)
Rename Brx.match_prefix_positions to Brx.bad_prefix_position.  The
function takes a regexp and a string not matching the regexp and
returns the index of the first character making the string not match
the regexp.

Make this function visible in boomerang as String.bad_index so the
user can use it and give an better error message (with line number and
character).

Fix harmony-export (a missing pop in the directory stack)