Parallel checking for non-mandatory contracts.  Unscientific running times:
  MODE         TIME
  -----------------
  (default)   48.79
  -unsafe     12.34
  -parallel   32.78

Mandatory checks are written with where! instead of where.  These are
NEVER run in parallel.

When running with -parallel, a workqueue is created for each module.
When running the module, any non-mandatory contracts encountered will
be forked to a separate thread to be executed in order.  If they fail,
they'll raise a top-level blame error and stop the system.  Any checks
encountered _within_ another check are run in-line, and not deferred
to the workqueue.

Unscientifically, uniProtV2 runs 26452 checks in parallel and 45829
in-line.  These counts aren't perfect---we're counting a check as
"in-line" when it's run from inside another check, as well.

The form

  test e = error

will cause e to be executed in-line the main thread---the error
condition may not come from a mandatory contract.  (This isn't
perfect---ideally, we'd be able to insert a "punctuation" in the
workqueue that we could wait for, and then ensure that all checks up
to that test were performed.  Currently, workqueues can't be waited
for without being stopped.)

Boomerang closures had to be slightly changed: they now take a
workqueue.  This prevents buggy behavior in tests.

The change necessitates adding -thread as an OCaml build option.  This
shouldn't be an issue, since all platforms should support threads.

There's plenty of room for improvement: workqueues could be shared
between modules, the workqueue implementation could be improved, and
error reporting could be more tightly integrated.  Workqueues also
need to be more carefully deallocated.  Finally, some analysis needs
to be done in terms of workqueue/GC interactions.  How long do checks
live in the workqueue?  Are pending checks delaying GC of what would
otherwise be garbage?